usecase_heat_waves/requirements.txt
streamlit
folium
geopandas
plotly
shapely
streamlit-folium
pillow
osmnx
tqdm
googletrans==3.1.0-alpha
xarray
rioxarray
meteostat==1.6.7
psycopg2
arcgis
dask[dataframe]